federated_sentiment_analysis/requirements.txt
pandas
numpy
fastapi
uvicorn
torch
requests
kafka-python
nltk
scikit-learn